Moved duck typing traits to "quack"
Closes https://github.com/PistonDevelopers/current/issues/77
